Fix 32bit compile error

+++ include/dim/util.h
@@ -71,7 +71,7 @@ inline int digits10(uint32_t val) {
 
 //===========================================================================
 // Round up to power of 2
-inline size_t pow2Ceil(size_t num) {
+inline size_t pow2Ceil(uint64_t num) {
 #if 0
     unsigned long k;
     _BitScanReverse64(&k, num);
